control_interface/requirements.txt
PyQt5
numpy
opencv-python
websocket-client
python-socketio[client]
dotenv
python-dotenv

torch
torchvision
Pillow
mediapipe
requests